% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoNW.R
\name{ontoNW}
\alias{ontoNW}
\title{ontoNW}
\usage{
ontoNW(onto, cutoff = 0.05, pdf = T, html = T, dynamic = T)
}
\arguments{
\item{onto}{data.frame. Includes an 'id' column with term id (e.g. G0:0008150), a 'type' colum (e.g. BP, KEGG), a 'name' column (e.g. biological_process), a 'genes' column with string of genes of the query belonging to the ontology term (e.g. IL1RL1, IL5RA, TP53), and a 'fdr' column with FDR values.}

\item{cutoff}{numeric. A fdr value as cutoff threshold to plot edges/nodes below cutoff}

\item{pdf}{boolean. Saves the static network as pdf.}

\item{html}{boolean. Saves the dynamic network as html. Applies only if dynamic=T.}

\item{dynamic}{boolean. Produce a dynamic network.}
}
\value{
nodes. nodesCut. Data frames with nodes id, pval, name, and type. Cutoff applies to nodesCut.

edges. edgesCut. Data frames with edges from node, to node, and weight (p-value). Cutoff applies to edgesCut.

net. netCut. igraph object. Cutoff applies to netCut.

DynNetwork. visNetwork object and htmlwidget. Cutoff applies.
}
\description{
\strong{Ontology Network}. Build gene-ontology relationship static and dynamic network.
}
\details{
Layout are lgl for static network and mds for dynamic network. Read igraph and visNetwork documentation.
}
\examples{
onto=data.frame(id  = c("GO:0002376", "GO:0032634")
                type= c("BP","BP")
                name= c("immune system process","interleukin-5 production")
                gene= c(c(IL5RA,SMPD3,CLC,IL1RL1,CYSLTR2,ALOX15,PIK3R6), c(IL5RA,IL1RL1))
                fdr = c(0.005, 0.00005)
                )
ontoNW(onto, cutoff=0.05, dynamic=T, pdf=F, html=T)       
}
\author{
Nathan Lemonnier \email{nathanael.lemonnier@gmail.com}
}
